+++ src/search-filter-api/src/patch.rs
@@ -109,7 +109,6 @@ mod tests {
             &get_product_service,
             &query_product_service,
             None,
-            None,
             &personalization_service,
         )
         .await
@@ -144,7 +143,6 @@ mod tests {
             &service,
             &get_product_service,
             &query_product_service,
-            None,
             None,
             &personalization_service,
         )
@@ -182,7 +180,6 @@ mod tests {
             &get_product_service,
             &query_product_service,
             None,
-            None,
             &personalization_service,
         )
         .await
@@ -214,7 +211,6 @@ mod tests {
             &service,
             &get_product_service,
             &query_product_service,
-            None,
             None,
             &personalization_service,
         )
@@ -260,7 +256,6 @@ mod tests {
             &service,
             &get_product_service,
             &query_product_service,
-            None,
             None,
             &personalization_service,
         )
@@ -314,7 +309,6 @@ mod tests {
             &service,
             &get_product_service,
             &query_product_service,
-            None,
             None,
             &personalization_service,
         )