+++ src/search-filter-api/src/post.rs
@@ -102,7 +102,6 @@ mod tests {
             &get_product_service,
             &query_product_service,
             None,
-            None,
             &personalization_service,
         )
         .await
@@ -133,7 +132,6 @@ mod tests {
             &service,
             &get_product_service,
             &query_product_service,
-            None,
             None,
             &personalization_service,
         )
@@ -169,7 +167,6 @@ mod tests {
             &service,
             &get_product_service,
             &query_product_service,
-            None,
             None,
             &personalization_service,
         )